refactor: split PreviewRegistry into focused files and add #Preview macros
Extract reusable fixture helpers into PreviewFixtures.swift (theme, mock data builders) so both PreviewRegistry and #Preview macros can share them. Split the 2066-line PreviewRegistry.swift into 5 extension files by domain (GitStatus, FileTree, AgentChat, Overlays, Chrome), leaving the main file at 531 lines with just the view(named:) dispatcher and Metal-dependent helpers. Add 10 new #Preview macros to non-Metal views: WhichKeyOverlay, MinibufferView, CompletionOverlay, PickerOverlay, NotificationCenterView, and all Settings tabs.
