+++ macos/Sources/Views/Settings/AppearanceSettingsView.swift
@@ -100,3 +100,23 @@ private struct ThemeSwatch: View {
             ))
     }
 }
+
+#Preview("Appearance Settings") {
+    let state = SettingsState()
+    state.isLoading = false
+    state.currentThemeName = "doom_one"
+    state.fontFamily = "Menlo"
+    state.fontSize = 13
+    state.fontWeight = "regular"
+    state.fontLigatures = true
+    state.themePreviews = [
+        Wire.ThemePreview(name: "Doom One", atom: "doom_one", editorBg: 0x282C34, editorFg: 0xBBC2CF, accent: 0x51AFEF),
+        Wire.ThemePreview(name: "Tokyo Night", atom: "tokyo_night", editorBg: 0x1A1B26, editorFg: 0xC0CAF5, accent: 0x7AA2F7),
+        Wire.ThemePreview(name: "Catppuccin Mocha", atom: "catppuccin_mocha", editorBg: 0x1E1E2E, editorFg: 0xCDD6F4, accent: 0x89B4FA),
+        Wire.ThemePreview(name: "Solarized Dark", atom: "solarized_dark", editorBg: 0x002B36, editorFg: 0x839496, accent: 0x268BD2),
+        Wire.ThemePreview(name: "Gruvbox Dark", atom: "gruvbox_dark", editorBg: 0x282828, editorFg: 0xEBDBB2, accent: 0xFE8019),
+        Wire.ThemePreview(name: "Nord", atom: "nord", editorBg: 0x2E3440, editorFg: 0xD8DEE9, accent: 0x88C0D0),
+    ]
+    return AppearanceSettingsView(state: state, encoder: nil)
+        .frame(width: 520, height: 360)
+}
